feat(user): Add CRUD operations to User and integrate JSON file support
In this commit, a new feature was added to the User service and controller that allows for numbers of CRUD operations. GET endpoint for retrieving all users and DELETE endpoint for deleting user by ID were added to `user.router.ts`. A method for reading data from User JSON file was added to `user.service.ts`. `dotenv` configuration was changed from `'.env'` to `'.env.local'` for better environment management. `fs` package was added to handle filesystem operations related to JSON files. For better security, `'.env.local'` was added to `.gitignore`.

Language used in error messages in `user.controller.ts` was also changed to French for better regional support.

These changes are necessary to allow the application to