fix(readme): Correct endpoints format in README.md
The commit adjusts the format of API endpoints in the README.md file. The ID placeholders in URLs for "DeleteUserById" and "getUserById" endpoints were changed from "<ID>" to ":ID" to depict the correct format that is followed in the code and hence, reduce potential confusion for users who try to use it.