rm(tsconfig): Simplify tsconfig.json file by removing comments
This commit removes the extensive comments in the tsconfig.json file, keeping only the necessary ones. The aim is to simplify the file, making it easier to read and manage. All the removed options were commented out and not in use, causing unnecessary visual clutter. The active options in the config remains unchanged.