feat(test + fix): Add getByID test in UserJSONService and User declaration in UserJSONService
In this commit, a unit test was added to the 'getById' function of the 'UserJSONService' to ensure its functionality. This was done to ensure that we handle situations where user ID does not exist correctly. Besides, a type declaration of a constant 'User' was added in 'UserJSONService'. It was done to improve code consistency and readability by making sure that each declared variable has a data type.